Implement get latest record logic

+++ src/main/scala/btc/wallet/models/requests/RecordRequester.scala
@@ -1,3 +1,3 @@
 package btc.wallet.models.requests
 
-final case class RecordRequester(datetime: String, amount: Int)
+final case class RecordRequester(datetime: String, amount: Float)
