+++ src/main/scala/btc/wallet/services/interfaces/IRecordService.scala
@@ -6,7 +6,7 @@ import btc.wallet.repositories.interfaces.IRecordRepository
 import scala.concurrent.Future
 
 trait IRecordService {
-  def saveRecord(dateTime: String, amount: Int): Future[RecordResponder]
+  def saveRecord(dateTime: String, amount: Float): Future[RecordResponder]
   def getHistory(startDate: String, endDate: String): Future[HistoryResponder]
   def recordRepository: IRecordRepository
 }
